+++ mobile/android/android-components/components/browser/session/src/test/java/mozilla/components/browser/session/storage/SnapshotSerializerTest.kt
@@ -7,6 +7,8 @@ package mozilla.components.browser.session.storage
 import mozilla.components.browser.session.Session
 import org.json.JSONObject
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertFalse
+import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
@@ -20,6 +22,7 @@ class SnapshotSerializerTest {
             source = Session.Source.ACTION_VIEW,
             id = "test-id").apply {
             title = "Hello World"
+            readerMode = true
         }
 
         val json = serializeSession(originalSession)
@@ -29,6 +32,7 @@ class SnapshotSerializerTest {
         assertEquals(Session.Source.ACTION_VIEW, restoredSession.source)
         assertEquals("test-id", restoredSession.id)
         assertEquals("Hello World", restoredSession.title)
+        assertTrue(restoredSession.readerMode)
     }
 
     @Test
@@ -45,5 +49,6 @@ class SnapshotSerializerTest {
         assertEquals("https://www.mozilla.org", restoredSession.url)
         assertEquals(Session.Source.ACTION_VIEW, restoredSession.source)
         assertEquals("test-id", restoredSession.id)
+        assertFalse(restoredSession.readerMode)
     }
 }