+++ mobile/android/android-components/components/feature/readerview/src/main/java/mozilla/components/feature/readerview/view/ReaderViewControlsBar.kt
@@ -15,8 +15,8 @@ import android.util.AttributeSet
 import android.view.View
 import android.widget.RadioGroup
 import mozilla.components.feature.readerview.R
-import mozilla.components.feature.readerview.ReaderViewFeature.Config.ColorScheme
-import mozilla.components.feature.readerview.ReaderViewFeature.Config.FontType
+import mozilla.components.feature.readerview.ReaderViewFeature.ColorScheme
+import mozilla.components.feature.readerview.ReaderViewFeature.FontType
 
 const val MAX_TEXT_SIZE = 9
 const val MIN_TEXT_SIZE = 1
@@ -54,7 +54,7 @@ class ReaderViewControlsBar @JvmOverloads constructor(
     override fun setFont(font: FontType) {
         val selected = when (font) {
             FontType.SERIF -> R.id.mozac_feature_readerview_font_serif
-            FontType.SANS_SERIF -> R.id.mozac_feature_readerview_font_sans_serif
+            FontType.SANSSERIF -> R.id.mozac_feature_readerview_font_sans_serif
         }
         fontGroup.check(selected)
     }
@@ -120,10 +120,11 @@ class ReaderViewControlsBar @JvmOverloads constructor(
         super.onFocusChanged(gainFocus, direction, previouslyFocusedRect)
     }
 
+    @Suppress("ComplexMethod")
     private fun bindViews() {
         fontGroup = applyCheckedListener(R.id.mozac_feature_readerview_font_group) { checkedId ->
             val fontType = when (checkedId) {
-                R.id.mozac_feature_readerview_font_sans_serif -> FontType.SANS_SERIF
+                R.id.mozac_feature_readerview_font_sans_serif -> FontType.SANSSERIF
                 R.id.mozac_feature_readerview_font_serif -> FontType.SERIF
                 else -> FontType.SERIF
             }
@@ -139,10 +140,10 @@ class ReaderViewControlsBar @JvmOverloads constructor(
             listener?.onColorSchemeChanged(colorSchemeChoice)
         }
         fontIncrementButton = applyClickListener(R.id.mozac_feature_readerview_font_size_increase) {
-            listener?.onFontSizeIncreased()
+            listener?.onFontSizeIncreased()?.let { setFontSize(it) }
         }
         fontDecrementButton = applyClickListener(R.id.mozac_feature_readerview_font_size_decrease) {
-            listener?.onFontSizeDecreased()
+            listener?.onFontSizeDecreased()?.let { setFontSize(it) }
         }
     }